+++ lib/pbio/include/pbsys/host.h
@@ -29,11 +29,17 @@ typedef enum {
     PBSYS_HOST_TRANSPORT_TYPE_USB,
 } pbsys_host_transport_type_t;
 
+/**
+ * Storage size for the hub name, including null terminator.
+ */
+#define PBSYS_HOST_HUB_NAME_SIZE (16)
+
 #if PBSYS_CONFIG_HOST
 
 void pbsys_host_debug_print(const char *data, size_t len);
 bool pbsys_host_get_event_buf(pbsys_host_transport_type_t transport, uint8_t **buf, uint32_t **len);
 void pbsys_host_get_hub_capabilities(uint8_t *buf, pbsys_host_transport_type_t transport);
+const char *pbsys_host_get_hub_name(void);
 uint32_t pbsys_host_read_characteristic(uint8_t service, uint16_t char_id, pbsys_host_transport_type_t transport, uint8_t *buf, uint32_t buf_size);
 void pbsys_host_init(void);
 bool pbsys_host_is_connected(void);